# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Restaurant API",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/dish": {
            "get": {
                "tags": [
                    "Dish"
                ],
                "summary": "Вывод списка блюд",
                "operationId": "2ee9f347006c28d6fafe539972012882",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Страница",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Записей на странице",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 20
                    },
                    {
                        "name": "name",
                        "in": "query",
                        "description": "Наименование",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Кофе"
                    },
                    {
                        "name": "composition",
                        "in": "query",
                        "description": "Состав",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "Вода, сахар,..."
                    },
                    {
                        "name": "category_id",
                        "in": "query",
                        "description": "Категория меню",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    },
                    {
                        "name": "sort_by",
                        "in": "query",
                        "description": "Поле для сортировки",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "name"
                    },
                    {
                        "name": "sort_dir",
                        "in": "query",
                        "description": "Направление сортировки",
                        "required": false,
                        "schema": {
                            "type": "string"
                        },
                        "example": "desc"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Кофе"
                                                    },
                                                    "image": {
                                                        "type": "string",
                                                        "example": "Images/Dish/S9EXsmeHwGPkewmzcm0f6T9RtLHiYjzjvEj9z1Tx.jpg"
                                                    },
                                                    "composition": {
                                                        "type": "string",
                                                        "example": "Вода, сахар,..."
                                                    },
                                                    "calories": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "122"
                                                    },
                                                    "price": {
                                                        "type": "number",
                                                        "format": "float",
                                                        "example": "68.5"
                                                    },
                                                    "category_id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "email has been requred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Dish"
                ],
                "summary": "Ввод нового блюда",
                "operationId": "9b66ff5a7cc36a7c5b55943ed1d03e4c",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Кофе"
                                    },
                                    "image": {
                                        "type": "string",
                                        "format": "binary"
                                    },
                                    "composition": {
                                        "type": "string",
                                        "example": "Вода, сахар,..."
                                    },
                                    "calories": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "122"
                                    },
                                    "price": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "68.5"
                                    },
                                    "category_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ok",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Кофе"
                                                },
                                                "image": {
                                                    "type": "string",
                                                    "example": "Images/Dish/S9EXsmeHwGPkewmzcm0f6T9RtLHiYjzjvEj9z1Tx.jpg"
                                                },
                                                "composition": {
                                                    "type": "string",
                                                    "example": "Вода, сахар,..."
                                                },
                                                "calories": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": "122"
                                                },
                                                "price": {
                                                    "type": "number",
                                                    "format": "float",
                                                    "example": "68.5"
                                                },
                                                "category_id": {
                                                    "type": "integer",
                                                    "example": 1
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Ошибка валидации",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "email has been requred"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Ошибка сервера",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Error": {
                                            "type": "boolean",
                                            "example": "true"
                                        },
                                        "Message": {
                                            "type": "string",
                                            "example": "Error text"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    },
...
}